requirements.txt
absl-py==1.0.0
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.8.2
blinker==1.4
boto3==1.23.8
botocore==1.26.8
Brlapi==0.7.0
cachetools==5.1.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.30
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
dnspython==1.16.0
easydict==1.9
entrypoints==0.3
filelock==3.7.0
fire==0.4.0
flake8==4.0.1
fonttools==4.33.3
future==0.18.2
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
h5py==3.7.0
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
imageio==2.19.2
IMDbPY==6.8
importlib-metadata==4.11.4
isort==4.3.21
jmespath==1.0.0
keyring==18.0.1
kiwisolver==1.4.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupSafe==1.1.0
matplotlib==3.5.2
mccabe==0.6.1
nemo-emblems==5.2.0
netaddr==0.7.19
netifaces==0.10.4
numpy==1.22.4
oauthlib==3.1.0
onboard==1.4.1
opencv-python==4.5.5.64
packaging==20.3
PAM==0.4.2
pandas==1.4.2
pexpect==4.6.0
Pillow==9.1.1
protobuf==3.20.1
psutil==5.5.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pyflakes==2.4.0
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-file==1.4.3
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
s3transfer==0.5.2
sahi==0.9.3
scipy==1.8.1
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
setproctitle==1.1.10
Shapely==1.8.2
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
systemd-python==234
tb-nightly==2.10.0a20220526
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminaltables==3.1.10
thop==0.0.31.post2005241907
tinycss2==1.0.2
tldextract==2.2.1
torch==1.11.0
torchreid==1.4.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.1.2
xkit==0.0.0
xlrd==1.1.0
yacs==0.1.8
yapf==0.32.0
yolov5==6.1.2
youtube-dl==2021.4.26
zipp==3.8.0
